file section done!

+++ controllers/admin.js
@@ -1,3 +1,4 @@
+const fileHelper = require("../util/file");
 const { validationResult } = require("express-validator");
 const Product = require("../models/product");
 
@@ -73,7 +74,15 @@ exports.postAddProduct = async (req, res, next) => {
 
 exports.postDeleteProduct = async (req, res, next) => {
   const productId = req.body.productId;
+  const product = await Product.findById(productId);
+  if (!product) {
+    return next(new Error("Product not found!"));
+  }
+
+  fileHelper.deleteFile(product.imageUrl);
+
   await Product.deleteOne({ _id: productId, userId: req.user._id });
+
   res.redirect("/admin/products");
 };
 
@@ -135,7 +144,10 @@ exports.postEditProduct = async (req, res, next) => {
   product.title = title;
   product.price = price;
   product.description = description;
-  if (image) product.imageUrl = image.path;
+  if (image) {
+    fileHelper.deleteFile(product.imageUrl);
+    product.imageUrl = image.path;
+  }
 
   await product.save();
   res.redirect("/admin/products");
